Fix first set of API tests
